requirements.txt
 functions-framework==3.*
  Pillow==10.1.0
  firebase-admin==6.2.0